DS_assignment1/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
sklearn2
numpy 
scipy
Seaborn